+++ src/action.ts
@@ -1,32 +1,11 @@
 'use server';
 
-import type { MemoProps } from '@/types/memo';
-
-export const fetchCreateMemo = async (memoData: Omit<MemoProps, 'id' | 'createdAt'>) => {
-  try {
-    const response = await fetch(`http://localhost:3000/api/memos`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(memoData),
-    });
-
-    if (!response.ok) {
-      throw new Error('메모 생성에 실패했습니다.');
-    }
-    return response.json();
-  } catch (error) {
-    console.error('메모 생성 중 오류 발생:', error);
-  }
-};
-
-export const fetchCreateCategory = async (category: {category: string}) => {
+export const fetchCreateCategory = async (category: { category: string }) => {
   try {
     const response = await fetch(`http://localhost:3000/api/categories`, {
       method: 'POST',
       headers: {
-        'Content-Type' : 'application/json',
+        'Content-Type': 'application/json',
       },
       body: JSON.stringify(category),
     });
@@ -35,7 +14,7 @@ export const fetchCreateCategory = async (category: {category: string}) => {
       throw new Error('메모 생성에 실패했습니다.');
     }
     return response.json();
-  } catch(error) {
+  } catch (error) {
     console.error('카테고리 생성 중 오류 발생:', error);
     throw new Error('메모 생성 중 오류가 발생했습니다.');
   }
